fix warning of comparison with signed number and unsigned number

+++ src/decode.cc
@@ -38,9 +38,9 @@ namespace swarm {
     return true;
   }
 
-  int DecoderMap::build_decoder_vector (NetDec * nd,
-                                        std::vector <Decoder *> *dec_vec,
-                                        std::vector <std::string> *dec_name) {
+  size_t DecoderMap::build_decoder_vector (NetDec * nd,
+                                           std::vector<Decoder *> *dec_vec,
+                                           std::vector<std::string> *dec_name) {
     // TODO(masa): need to check contents of dec_vec, dec_name
     const size_t len = DecoderMap::protocol_decoder_map_.size ();
     size_t i = 0;
